Added authorization check for delete methods. Only level 1 users can delete resources

+++ app/Http/Controllers/CustomersController.php
@@ -126,7 +126,7 @@ class CustomersController extends Controller
      */
     public function deleteCustomer($id)
     {
-        // $this->authorize('delete', $id);
+        $this->authorize('delete', $id);
         return Customer::destroy($id);
     }
 }
